Fix bug, clone search filters for "load more" func
Because both; search and "load more" depend on the same search filters,
we have to fix the filters of load more function so that filters will
not be changed at run time if the user change them after the search is
invoked with its filters. That means once a search is invoked, then
"load more" must use the same set filters of that search. Changes to
the filters must not affect "load more" but can take effect on new
search.