Add icon as button to remove `recipient` filter
Add label with icon to be used as button to remove current `recipient`
filter of outbounds search.

Also refactor code to move the removal action of `recipient` search
filter field from the modal dialog to the main app window, so the remove
button is inlined next to the `recipient` filter field.