Display loading animation during loading of images
Use `SwingWorker` to load images synchronously without blocking the GUI.

Also use `ScheduledThreadPoolExecutor` to delay the display of the
loading animation for brief amount of time. This delay is useful in case
there are no images available, hence the method `done()` of the
`SwingWorker` returns quickly and shows the feedback icon of "no image
available". In that fast response no need to display animation during,
but just jump to the no image feedback since it takes brief time.