Display the last selected image in item image form
In item form for adding and editing items; set the image form to display
the last selected image after selecting images through the file browser.

That is different from opening the form, where for opening; the app
displays the default image which may not be necessarily the last
selected image.