Display image in place of removed image
After removing an image, replace the removed image with the closest
image. If the removed image was at the end (was having the last
position), then display the image that is before it (was before it
before removal). If the removed image was at any place other than the
end, then display the image that is assigned the position of the removed
image.