Rename database attributes and remove backticks
Rename database attributes for better semantic meaning. Remove the
backticks that wrap most attribute names in SQL queries, because JDBC
change from MariaDB to Apache derby which does not accept the wrapping
backticks.

Most if not all of the queries are erroneous against Apache derby
syntax. Next commits are about to fix that to accommodate the change to
Apache derby.