Bug fix: remove source properly
If the "Inbound ID" filter is selected, clicking the remove button for
the "Source" filter has no effect which is intended. However, if the
"Inbound ID" filter is deselected, clicking the remove button for the
"Source" filter should have its expected effect. The bug is that in the
latter case, the removal does not function as intended, the "Source"
object should be nulled.