Make updating recipient image using old image name
This has the advantage that the program doesn't need to interact with
the database. That is it does not remove database record and insert a
new one for the image name.

The update occurs only in the file system — the old image file is
removed, and the new image file is moved to the application directory
using the same name as the old one, which already matches the
database record.